The user can create and watch feeds now.
There is a page named feeds. If the user is logged in he can view older feeds and create new feeds. The page shows 2(shoul be 20) feeds at the time. There are buttons Newer and Older with wich the user can view all feeds. The application is ready to use socket.io.